requirements.txt
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0

# Computer Vision
opencv-python>=4.6.0
albumentations>=1.3.0
timm>=0.9.0

# Experiment Tracking and Logging
wandb>=0.15.0
tensorboard>=2.10.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data Processing
tqdm>=4.64.0
h5py>=3.7.0

# Deployment and API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
pre-commit>=2.20.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.0.0 